Added Precision Div
Added a precision div for users to enter the level of precision they
want. It works pretty well, but it broke down on the irrational answer
square of 2. (Type in the numbers for x^2 - 2) It looks like perhaps
it's not storing enough digits in Xn, but delta continues to get
smaller. However, it's not tripping the exit condition for the loop, so
it goes through it 10,000 times before incorrectly declaring it doesn't
converge.