Made it a little prettier.
i added a clear button along with an insert button so we don't have to
type everything into the text boxes every single time we test.  I
commented out the code where the .keyup was used because the variables
weren't populating properly with a button because they were originally
populated when a keyup occured.