Added lots more styling and functionality
Much Stuff... Very wow....
Note:  When Red X button is pressed the variables clear along with input
boxes but can't seem to make the fade-out functionality work for the
second jumbotron.  Slider needs to be created for precision so that we
don't have to enter a decimal answer.  Slider also needs some styling...
Possibly need to add an info button so that user's know how to use
program.  variable is set for the fade in so that it can only fade in
once unless the red X button is pressed.