Resizes(kind of)
so i made a div to center the graph and make everything a little nicer
now.
the only issue is when the screen is resized it glitches a little.  it
seems to happen when you scroll in and out.  the issue in console is:
Cannot read property 'removeChild' of null.
I believe this is caused by the scaling of the board that is forced by
css.  help?

Also the graph size is changed