Blingified Newton's Method
The graph can be resized in real time
The graph color can change in real time
i added some coloring to the div's for each part of the app so its more
defined.  I think it looks pretty good but on the other hand its 12:30
am so everything seems awesome at this point. lol